refactor + changes to aipopover
