+++ src/components/editor/index.tsx
@@ -35,7 +35,9 @@ import { useRoom } from "liveblocks.config";
 import { useEffect, useRef, useState } from "react";
 import * as Y from "yjs";
 // import { CommentFormattingToolbarButton } from "@/components/Editor/CustomBlocks/Comment";
-import AiPopover, { AiPopoverPropsRect } from "@/components/editor/ai-popover";
+import AiPopover, {
+  AiPopoverPropsRect,
+} from "@/components/editor/custom/ai/popover";
 
 export default function Editor({
   canEdit,
